Added computed value to Action Button for holding css class logic
Computed values like the name implies are values in component that are derived and calculated from other data sources such as data or props.
So when the data they use changess, the result value is recalculaated and the component re-rendered if it was dependant on it.

Like with event handlers you pass the function name to the html template without invoking it, vue itself will rerun it when it's input data changes.
However unlike event handlers this is mostly used to simply display a value, much like stuff in data section of a component.