Stubbed out font-awesome-icon component that was causing warnings in unit tests.
Stubbing out means telling the test suite to just replace it with a ..well stub, and not treat it as anything important.
This needed to be done anywhere where SubNav is being rendered for testing which included not only tests for it, but
also tests for parent component, as this renders children as well.

adding configuration objext as secondary argument to render function allows to set up a list of components to stub.